packages/boto3/requirements.txt
boto3==1.24.46
python-dotenv==0.20.0
